refactor classes into less files to avoid circular imports
