+++ python/hsfs/constructor/on_demand_feature_group_alias.py
@@ -16,13 +16,13 @@
 
 import humps
 
-from hsfs import on_demand_feature_group as on_demand_fg
+from hsfs import feature_group
 
 
 class OnDemandFeatureGroupAlias:
     def __init__(self, on_demand_feature_group, alias):
         self._on_demand_feature_group = (
-            on_demand_fg.OnDemandFeatureGroup.from_response_json(
+            feature_group.OnDemandFeatureGroup.from_response_json(
                 on_demand_feature_group
             )
         )
