+++ python/hsfs/feature_group.py
@@ -21,12 +21,113 @@ import pandas as pd
 import numpy as np
 from typing import Optional, Union, Any, Dict, List, TypeVar
 
-from hsfs import util, engine, feature
-from hsfs.core import feature_group_engine, statistics_engine, feature_group_base
+from hsfs import util, engine, feature, storage_connector as sc
+from hsfs.core import (
+    feature_group_engine,
+    statistics_engine,
+    feature_group_base_engine,
+    on_demand_feature_group_engine,
+)
 from hsfs.statistics_config import StatisticsConfig
+from hsfs.constructor import query
 
 
-class FeatureGroup(feature_group_base.FeatureGroupBase):
+class FeatureGroupBase:
+    def __init__(self, featurestore_id):
+        self._feature_group_base_engine = (
+            feature_group_base_engine.FeatureGroupBaseEngine(featurestore_id)
+        )
+
+    def delete(self):
+        """Drop the entire feature group along with its feature data.
+
+        !!! danger "Potentially dangerous operation"
+            This operation drops all metadata associated with **this version** of the
+            feature group **and** all the feature data in offline and online storage
+            associated with it.
+
+        # Raises
+            `RestAPIError`.
+        """
+        self._feature_group_engine.delete(self)
+
+    def select_all(self):
+        """Select all features in the feature group and return a query object.
+
+        The query can be used to construct joins of feature groups or create a
+        training dataset immediately.
+
+        # Returns
+            `Query`. A query object with all features of the feature group.
+        """
+        return query.Query(
+            self._feature_store_name, self._feature_store_id, self, self._features
+        )
+
+    def select(self, features=[]):
+        """Select a subset of features of the feature group and return a query object.
+
+        The query can be used to construct joins of feature groups or create a training
+        dataset with a subset of features of the feature group.
+
+        # Arguments
+            features: list, optional. A list of `Feature` objects or feature names as
+                strings to be selected, defaults to [].
+
+        # Returns
+            `Query`: A query object with the selected features of the feature group.
+        """
+        return query.Query(
+            self._feature_store_name, self._feature_store_id, self, features
+        )
+
+    def add_tag(self, name: str, value: str = None):
+        """Attach a name/value tag to a feature group.
+
+        A tag can consist of a name only or a name/value pair. Tag names are
+        unique identifiers.
+
+        # Arguments
+            name: Name of the tag to be added.
+            value: Value of the tag to be added, defaults to `None`.
+
+        # Raises
+            `RestAPIError`.
+        """
+        self._feature_group_base_engine.add_tag(self, name, value)
+
+    def delete_tag(self, name: str):
+        """Delete a tag from a feature group.
+
+        Tag names are unique identifiers.
+
+        # Arguments
+            name: Name of the tag to be removed.
+
+        # Raises
+            `RestAPIError`.
+        """
+        self._feature_group_base_engine.delete_tag(self, name)
+
+    def get_tag(self, name: str = None):
+        """Get the tags of a feature group.
+
+        Tag names are unique identifiers. Returns all tags if no tag name is
+        specified.
+
+        # Arguments
+            name: Name of the tag to get, defaults to `None`.
+
+        # Returns
+            `list[Tag]`. List of tags as name/value pairs.
+
+        # Raises
+            `RestAPIError`.
+        """
+        return self._feature_group_base_engine.get_tags(self, name)
+
+
+class FeatureGroup(FeatureGroupBase):
     CACHED_FEATURE_GROUP = "CACHED_FEATURE_GROUP"
     ENTITY_TYPE = "featuregroups"
 
@@ -638,3 +739,165 @@ class FeatureGroup(feature_group_base.FeatureGroupBase):
             return self.statistics
         else:
             return self._statistics_engine.get(self, commit_time)
+
+
+class OnDemandFeatureGroup(FeatureGroupBase):
+    ON_DEMAND_FEATURE_GROUP = "ON_DEMAND_FEATURE_GROUP"
+
+    def __init__(
+        self,
+        query,
+        storage_connector,
+        name=None,
+        version=None,
+        description=None,
+        featurestore_id=None,
+        featurestore_name=None,
+        created=None,
+        creator=None,
+        id=None,
+        features=None,
+        jobs=None,
+        desc_stats_enabled=None,
+        feat_corr_enabled=None,
+        feat_hist_enabled=None,
+        cluster_analysis_enabled=None,
+        statistic_columns=None,
+        statistics_config=None,
+    ):
+        super().__init__(featurestore_id)
+
+        self._feature_store_id = featurestore_id
+        self._feature_store_name = featurestore_name
+        self._description = description
+        self._created = created
+        self._creator = creator
+        self._version = version
+        self._name = name
+        self._query = query
+        self._id = id
+        self._jobs = jobs
+        self._desc_stats_enabled = desc_stats_enabled
+        self._feat_corr_enabled = feat_corr_enabled
+        self._feat_hist_enabled = feat_hist_enabled
+        self._statistic_columns = statistic_columns
+
+        self._feature_group_engine = (
+            on_demand_feature_group_engine.OnDemandFeatureGroupEngine(featurestore_id)
+        )
+
+        if self._id:
+            # Got from Hopsworks, deserialize features and storage connector
+            self._features = [
+                feature.Feature.from_response_json(feat) for feat in features
+            ]
+        else:
+            self._features = features
+
+        if storage_connector is not None and isinstance(storage_connector, dict):
+            self._storage_connector = sc.StorageConnector.from_response_json(
+                storage_connector
+            )
+        else:
+            self._storage_connector = storage_connector
+
+    def save(self):
+        self._feature_group_engine.save(self)
+
+    def read(self, dataframe_type="default"):
+        """Get the feature group as a DataFrame."""
+        engine.get_instance().set_job_group(
+            "Fetching Feature group",
+            "Getting feature group: {} from the featurestore {}".format(
+                self._name, self._feature_store_name
+            ),
+        )
+        return self.select_all().read(dataframe_type=dataframe_type)
+
+    def show(self, n):
+        """Show the first n rows of the feature group."""
+        engine.get_instance().set_job_group(
+            "Fetching Feature group",
+            "Getting feature group: {} from the featurestore {}".format(
+                self._name, self._feature_store_name
+            ),
+        )
+        return self.select_all().show(n)
+
+    @classmethod
+    def from_response_json(cls, json_dict):
+        json_decamelized = humps.decamelize(json_dict)
+        if "type" in json_decamelized:
+            _ = json_decamelized.pop("type")
+        return cls(**json_decamelized)
+
+    def update_from_response_json(self, json_dict):
+        json_decamelized = humps.decamelize(json_dict)
+        if "type" in json_decamelized:
+            _ = json_decamelized.pop("type")
+        self.__init__(**json_decamelized)
+        return self
+
+    def json(self):
+        return json.dumps(self, cls=util.FeatureStoreEncoder)
+
+    def to_dict(self):
+        return {
+            "id": self._id,
+            "name": self._name,
+            "description": self._description,
+            "version": self._version,
+            "features": self._features,
+            "featurestoreId": self._feature_store_id,
+            "query": self._query,
+            "storageConnector": self._storage_connector.to_dict(),
+            "type": "onDemandFeaturegroupDTO",
+        }
+
+    @property
+    def id(self):
+        return self._id
+
+    @property
+    def name(self):
+        return self._name
+
+    @property
+    def version(self):
+        return self._version
+
+    @property
+    def description(self):
+        return self._description
+
+    @property
+    def features(self):
+        return self._features
+
+    @property
+    def query(self):
+        return self._query
+
+    @property
+    def storage_connector(self):
+        return self._storage_connector
+
+    @property
+    def creator(self):
+        return self._creator
+
+    @property
+    def created(self):
+        return self._created
+
+    @version.setter
+    def version(self, version):
+        self._version = version
+
+    @description.setter
+    def description(self, new_description):
+        self._description = new_description
+
+    @features.setter
+    def features(self, new_features):
+        self._features = new_features
